+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased (1.0.1)
 
+* Added utilities for converting Capstone operands to assembly syntax (x86)
+* Added a utility to get the symbolic expression for an operand (x86)
+
 # 1.0.0
 
 * Initial version